% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_used_packages}
\alias{get_used_packages}
\title{Get used packages from parsed R code tibble}
\usage{
get_used_packages(parsed_tbl)
}
\arguments{
\item{parsed_tbl}{parsed R code tibble from from \code{parser()}}
}
\value{
tibble with used packages from R code
}
\description{
Get used packages from parsed R code tibble
}
\examples{
get_used_functions(parser(file = "tests/testthat/ref/01-iris-count.R"))
}
