% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser}
\alias{parser}
\title{Parse R code}
\usage{
parser(file = "", text = NULL)
}
\arguments{
\item{file}{string a filename or URL to read expressions from}

\item{text}{optional character vector text to parse}
}
\value{
tibble with parsed data
}
\description{
Parse R code
}
\examples{
parser(text = "function(a = 1, b = 2) { \n  a + b\n}\n")
}
